chore: remove usePublishedRemoteUsers and implement useRemoteUsers
