+++ packages/agora-rtc-react/test/hook/useRemoteUsers.test.ts
@@ -8,6 +8,9 @@ const setUp = (client?: IAgoraRTCClient | null) => renderHook(() => useRemoteUse
 
 describe("useRemoteUsers", () => {
   const userList = [{ uid: "1", hasAudio: true, hasVideo: true }];
+  const user1 = { uid: "1", hasAudio: true, hasVideo: true };
+  const user2 = { uid: "1", hasAudio: true, hasVideo: false };
+  const user3 = { uid: "1", hasAudio: false, hasVideo: true };
 
   test("should return user list when the provided client", () => {
     const client = FakeRTCClient.create();
@@ -37,5 +40,29 @@ describe("useRemoteUsers", () => {
       dispatchRTCEvent(client, "user-left");
     });
     expect(result.current.length).toBe(0);
+  });
+
+  test("should return user list when user-published", async () => {
+    const client = FakeRTCClient.create();
+    const { result } = setUp(client);
+    act(() => {
+      (client.remoteUsers as IAgoraRTCRemoteUser[]) = [user1];
+      dispatchRTCEvent(client, "user-published");
+    });
+    await waitFor(() => {
+      expect(result.current.length).toBe(client.remoteUsers.length);
+    });
+  });
+
+  test("should return user list when user-unpublished", async () => {
+    const client = FakeRTCClient.create();
+    const { result } = setUp(client);
+    act(() => {
+      (client.remoteUsers as IAgoraRTCRemoteUser[]) = [user1, user2, user3];
+      dispatchRTCEvent(client, "user-unpublished");
+    });
+    await waitFor(() => {
+      expect(result.current.length).toBe(client.remoteUsers.length);
+    });
   });
 });
